Specify content type for uploaded files in S3, where one could be determined.
This requires a change to RobustHTTPClient to accept a content type parameter for an uploaded file, and for simplicity's sake I've opted to include a copy of that class in this repo. For a proper upstream PR, the changes to RobustHTTPClientEx in this commit would need to be made in the `org.jenkins-ci.plugins.apache-httpcomponents-client-4-api` project instead.

As far as determining the appropriate content for a given file, at first `java.nio.Files#probeContentType()` is called to see if that returns a non-null content type. If unsuccessful, `java.net.URLConnection#guessContentTypeFromName()` is used to pick a content type based on the file extension. For each file in the list of artifacts to be uploaded, the content type (if any) is recorded in a separate map. Additionally, the content type is passed in to the S3BlobStore method which calculates the presigned URL since the content-type header is part of the signed URL and needs to be known before generating the URL. Next, during the loop which uploads the artifacts to S3, the recorded content types for each file (again, if any was able to be determined for a given file) are retrieved and passed in to the RobustHTTPClient's `uploadFile` call.

In S3BlobStore, we need to use `GeneratePresignedUrlRequest` to generate the presigned URL as it allows us to specify the "Content-Type" header in the generated URL. The existing `AmazonS3ClientBuilder` does not provide this option.

The existing RobustHTTPClient `uploadFile(File, URL, TaskListener)` method signature is preserved, and changed to call the new `uploadFile(File, String, URL, TaskListener)` which includes an additional parameter for specifying the content type. If this is non-null, the "Content-Type" HTTP header will be set accordingly. This is necessary, or the presigned URL which included the "Content-Type" header, and the `GET` request used to perform the upload do not match and will throw an AWS exception complaining that "The request signature we calculated does not match the signature you provided."

Signed-off-by: Robin Verduijn <robinverduijn.github@gmail.com>